Remove unused imports
Because if they ae not used, why import them?
Don't worry, I did not do that by hand, it's a feature my editor does for me.

For the `tessilator.py` itself, it's a question of we want to do that because that means that also would not get added to `__all__`. Personally, I'd prefer if users import them for where they comes from and not from tessilator itself, because then you end up in a similar situation as you do with "*" imports - you get a bunch of functions, but don't really know where they come from.
 but there might valid reasons to do it this was. It's just a question of design, so we might want to discuss that.