Fix some minor Python style issues
such as use of f like `f"string"` when there is nothing in the string to be formatted, etc.
Nothing of this is wrong or will change the execution of the program, I'm mostly fixing it because my editor highlights it as red and then I might miss the real errors.