simply structure in all_sources_sector
Before this fix, on branch used `res_table, names, formats = create_table_template()`
and the other branch used `res_table = create_table_template()`.
Only one of those is correct. So, I changed the structure, such that the code is the same every time and does not go out of sync.