combine all the `cutout_X` functions into a single func
Except for the error messages printed out, these functions were essentially identical (and probably developed by copying and pasting from one to the other). I combined them back into one function, so that we don't have to maintain them separately. In fact, one of them had an error in it (I think a typo in a variable name) that would have made it fail, if it was ever used (which shows our test coverage was insufficient).