Change minimum Python version to 3.11
The commit contains another change:
Moving from setup.py entirely to pyproject.toml.
That change is not required, but the reasons I'm doing it here is quite stupid:
I did not know how to set the minimum Python version in setup.cfg, so I copied it from the astropy setup and that uses pyproject.toml. So, because I'm lazy, I did that change as well. `pyproject.toml` is considered the more modern version of setting up projects, so the change (if I've transferred all the info correctly) certainly does not hurt.