Use context managers instead of setting the global rc_param state
There were two locations in the tessilator, where the tessilator changes the global state of the matplotlib defaults (the font size).
When using the tessilator in the Python program (as opposed to a command line tool that exists after the plots are done), then any other plot made after calling the tessilator will also have those settings, if I like it or not.
So, instead, this commit replaces the change of the global state with a context that changes the fontsize only for that one plot.

Also included in the commit are a few unrelated typos that I found while looking through the files, and a simplifications of a loop, that I stumbled upon. None of them change the behavior, and I just figured I might as well fix them if I happen to catch them serendipously.