Fix legend formatting for newer versions of matplotlib
For the current matplotlib (I believe from 3.9.0 on) legend no longer has a `legendHandles` attribute, it is now named differently. This leads to a an exception in matplotlib, which will cause `createplot` to not create a plot, and instead raise an exception. Unfortunately, the user never sees that exception: The exception bubbles up through `full_run_lc` which in turn is called by `one_source_cutout` and in `one_source_cutout` there is a blanket statement that catches all exceptions and simply writes a single line tothe logfile ("Error occurred when processing {file_in}. Trying next target.").
Thus, we never know that there is a problem and that that problem comes from a change in matplotlib.

I found it when running the tessilator tests with remote data, which fails because the plot is not created. It still took me a long time to dig down to WHY the plot is not created.

This commit fixes the matplotlib problem using a different way to set the size of the scatter dots in the legend that is hopefully more portable between version as it does not rely on an attribute of method that is marked private.